+++ releasenotes.md
@@ -1,3 +1,4 @@
 Whats New:
 ----------------------
 We added the option to select between 2D and 3D for PubChem image. Go to settings and select the desirable format!
+You can now add a trading name along with INCI name in ingredients